+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.4.1
+
+- [Fixing VALUES-clause update query naming](https://github.com/zio/zio-quill/pull/2595)
+
+# 4.4.0
+
+- [UPDATE with VALUES optimization for Postgres. Various macro refactoring.](https://github.com/zio/zio-quill/pull/2571)
+
 # 4.3.0
 
 - [Values clause batch insert](https://github.com/zio/zio-quill/pull/2565)